[Heartbeat] Cog model restructure
Restructure cog `Heartbeat` according to SFUAnime/Ren#608. No functional
changes.

Constants are moved to `constants.py`.
